Connector option to hide indirect (ORG_FLAG not empty)
This change introduces a new configuration option to hide every indirect ACTIVITGROUP
(identified by the ORG_FLAG tag in the ri:ACTIVITYGROUPS field being not empty).
This was introduced because the indirect and complex SAP role hierarchies are not important
for the user administration in midpoint. Managing those is quiet complicated because the easiest
way to manage the association to activigroups is the ACTIVITYGROUPS.AGR_NAME field on the user.
To ignore indirect activitygroups inside midPoint you need complicated mappings which are difficult
to test. With this change its quite easy to hide them from midPoint.

The feature is disabled on default.

The actual filtering will be done in the connector and not in the SAP query.